Fixed Windows support.
After rainbow didn't work on some of my Windows based ruby environments I
started investigating. The major reason for the problem was that some of of the
environments returned mingw32 instead of mswin32 because they are built in a
minimal unix compatibility framework. Apart of that I think depending on the
architecture the Windows system runs upon isn't really necessary most of the
time, so i got rid of the 32 part. Finally, even though it seems JRuby currently
doesn't seem to provide an windows ANSI library, JRuby's RUBY_PLATFORM always
returns 'java'. Therefore the use of Config::CONFIG['host_os'] is highly
recommended to be compatible with a wide range of Ruby implementations.